third_party/WebKit/LayoutTests/http/tests/security/xss-DENIED-window-open-parent-expected.txt

CONSOLE ERROR: line 2: Unsafe JavaScript attempt to initiate navigation for frame with URL 'http://127.0.0.1:8000/security/xss-DENIED-window-open-parent.html' from frame with URL 'http://localhost:8080/security/resources/xss-DENIED-window-open-parent-attacker.html'. The frame attempting navigation is targeting its top-level window, but is neither same-origin with its target nor is it processing a user gesture. See https://www.chromestatus.com/features/5851021045661696.

This test passes if there is no alert dialog.
